fix: clone repo
