+++ src/adaos/sdk/skill_service.py
@@ -8,51 +8,21 @@ import importlib.util
 from adaos.core.i18n import _
 from adaos.sdk.context import SKILLS_DIR, TEMPLATES_DIR, MONOREPO_URL, get_current_skill_path, set_current_skill, current_skill_name, current_skill_path
 from adaos.db.sqlite import add_or_update_skill, update_skill_version, list_skills, set_installed_flag
+from adaos.utils.git_utils import _ensure_repo
 
 
 def _skill_subdir(skill_name: str) -> str:
     return skill_name
 
 
-def _ensure_repo() -> tuple[Repo, bool]:
-    os.makedirs(Path(SKILLS_DIR), exist_ok=True)
-    git_dir = os.path.join(Path(SKILLS_DIR), ".git")
-
-    if not os.path.exists(git_dir):
-        print(f"[cyan]{_('repo.clone')} {MONOREPO_URL}[/cyan]")
-        repo = Repo.clone_from(MONOREPO_URL, SKILLS_DIR)
-        repo.git.config("index.version", "2")
-        repo.git.sparse_checkout("init", "--cone")
-        return repo, True
-
-    repo = Repo(Path(SKILLS_DIR))
-
-    try:
-        current_index_ver = repo.git.config("index.version")
-    except:
-        current_index_ver = "3"
-
-    if current_index_ver != "2":
-        print(f"[yellow]{_('repo.rebuild_index')}[/yellow]")
-        repo.git.config("index.version", "2")
-        if repo.head.is_valid():
-            repo.git.reset("--mixed")
-
-    return repo, False
-
-
-def _sync_sparse_checkout(repo: Repo):
+def _sync_sparse_checkout(repo: Repo, installed=[]):
     """Пересобираем sparse-checkout из всех установленных навыков"""
-    installed = [s["name"] for s in list_skills() if s.get("installed", 1)]
+    installed = [s["name"] for s in list_skills() if s.get("installed", 1)] + installed
     repo.git.sparse_checkout("set", *installed)
-    """ if installed:
-        repo.git.sparse_checkout("set", *installed)
-    else:
-        repo.git.sparse_checkout("disable") """
 
 
 def create_skill(skill_name: str, template_name: str = "basic") -> str:
-    repo, is_initial = _ensure_repo()
+    repo = _ensure_repo()
     skill_subdir = _skill_subdir(skill_name)
     skill_path = os.path.join(Path(SKILLS_DIR), skill_subdir)
 
@@ -95,7 +65,7 @@ def create_skill(skill_name: str, template_name: str = "basic") -> str:
 
 
 def push_skill(skill_name, message: str = None) -> str:
-    repo, is_initial = _ensure_repo()
+    repo = _ensure_repo()
     _sync_sparse_checkout(repo)
     repo.git.add(skill_name)
 
@@ -108,11 +78,10 @@ def push_skill(skill_name, message: str = None) -> str:
 
 
 def pull_skill(skill_name: str) -> str:
-    repo, is_initial = _ensure_repo()
+    repo = _ensure_repo()
     set_installed_flag(skill_name, installed=1)
-    _sync_sparse_checkout(repo)
-    if not is_initial:
-        repo.remotes.origin.pull()
+    _sync_sparse_checkout(repo, installed=[skill_name])
+    repo.remotes.origin.pull()
 
     yaml_path = os.path.join(Path(SKILLS_DIR), skill_name, "skill.yaml")
     version = "unknown"
@@ -126,7 +95,7 @@ def pull_skill(skill_name: str) -> str:
 
 
 def update_skill() -> str:
-    repo, is_initial = _ensure_repo()
+    repo = _ensure_repo()
     _sync_sparse_checkout(repo)
     repo.remotes.origin.pull()
 
@@ -147,7 +116,7 @@ def install_skill(skill_name: str) -> str:
 
 def uninstall_skill(skill_name: str) -> str:
     """Удаляет навык у пользователя (ставим installed=0 и пересобираем sparse-checkout)"""
-    repo, is_initial = _ensure_repo()
+    repo = _ensure_repo()
     set_installed_flag(skill_name, installed=0)
     _sync_sparse_checkout(repo)
     return f"[green]{_('skill.uninstalled', skill_name=skill_name)}[/green]"
