+++ src/adaos/utils/git_utils.py
@@ -2,10 +2,37 @@ import os
 from pathlib import Path
 from git import Repo
 from dotenv import load_dotenv, find_dotenv
-from adaos.sdk.context import SKILLS_DIR
+from adaos.sdk.context import SKILLS_DIR, MONOREPO_URL
+from adaos.core.i18n import _
 
 GIT_USER = os.getenv("GIT_USER", "adaos")
 GIT_EMAIL = os.getenv("GIT_EMAIL", "adaos@local")
+
+
+def clone_git_repo() -> Repo:
+    os.makedirs(Path(SKILLS_DIR), exist_ok=True)
+    git_dir = os.path.join(Path(SKILLS_DIR), ".git")
+
+    if not os.path.exists(git_dir):
+        print(f"[cyan]{_('repo.clone')} {MONOREPO_URL}[/cyan]")
+        repo = Repo.clone_from(MONOREPO_URL, SKILLS_DIR)
+        repo.git.config("index.version", "2")
+        repo.git.sparse_checkout("init", "--cone")
+        return repo, True
+
+    repo = Repo(Path(SKILLS_DIR))
+
+    try:
+        current_index_ver = repo.git.config("index.version")
+    except:
+        current_index_ver = "3"
+
+    if current_index_ver != "2":
+        print(f"[yellow]{_('repo.rebuild_index')}[/yellow]")
+        repo.git.config("index.version", "2")
+        if repo.head.is_valid():
+            repo.git.reset("--mixed")
+    return repo
 
 
 def init_git_repo() -> Repo:
@@ -62,3 +89,15 @@ def rollback_last_commit():
     last_commit = repo.head.commit
     repo.git.reset("--hard", "HEAD~1")
     print(f"[GIT] Откат на коммит {last_commit.hexsha[:7]}")
+
+
+def _ensure_repo() -> tuple[Repo, bool]:
+    os.makedirs(Path(SKILLS_DIR), exist_ok=True)
+    git_dir = os.path.join(Path(SKILLS_DIR), ".git")
+
+    if not os.path.exists(git_dir):
+        repo = clone_git_repo()
+    else:
+        repo = Repo(Path(SKILLS_DIR))
+
+    return repo
