+++ src/adaos/utils/setup_env.py
@@ -4,7 +4,7 @@ import urllib.request
 import shutil
 from pathlib import Path
 from adaos.db.sqlite import init_db
-from adaos.utils.git_utils import init_git_repo
+from adaos.utils.git_utils import _ensure_repo
 from adaos.sdk.context import BASE_DIR, DB_PATH
 
 MODELS_DIR = f"{BASE_DIR}/models"
@@ -37,11 +37,12 @@ def prepare_environment():
     if not env_file.exists() and env_sample.exists():
         shutil.copy(env_sample, env_file)
         print(f"[AdaOS] Created .env file at {env_file}")
-    # Git репозиторий навыков
-    init_git_repo()
 
     # БД
     init_db()
+
+    # Git репозиторий навыков
+    _ensure_repo()
 
     # Модель wake-word
     download_vosk_model()
